separate groupMe logic into separate class
it's not functional (yet) but works
